First commit.
CLIGames is intended to be a library of games
for the command line. The idea is to have games
implemented and developed in different languages
in order to explore and learn.

Hangman game implemented in C and Python. Python
version will be deprecated soon.

Wordsearch implemented in Python3

lib folder contains txt files with list of words
used by the games.

TODO:
	Translate wordsearch to english (comments and code)
	is at the moment in spanish.